updating sphere reader
