+++ EpmPython/read.py
@@ -1,4 +1,3 @@
-from base_state import BaseState
 import sphere
 import h5py
 import numpy as np
@@ -9,7 +8,55 @@ env = os.environ.copy()
 sys.path.append(env['HOME']+'/quicc-github/QuICC/Python/')
 #get it from crossover_master_release
 #from quicc.projection.shell_energy import ortho_pol_q, ortho_pol_s, ortho_tor
-import integrateWorland as wor 
+#import integrateWorland as wor 
+
+class BaseState:
+    
+    def __init__(self, filename, geometry, file_type='QuICC'):
+        
+        # PhysicalState('state0000.hdf5',geometry='Sphere'
+        fin = h5py.File(filename, 'r')
+        self.fin = fin
+
+        
+        attrs = list(self.fin.attrs.keys())
+       
+        if len(attrs) > 2 : 
+            if attrs[2] == "Version" and file_type.lower()!= 'quicc':
+                raise RunTimeError("maybe your file is EPM")
+
+        if attrs[1] == "version" and file_type.lower()!= 'epm':
+            raise RunTimeError("maybe your file is QuICC")
+                            
+        # TODO: distinguish between EPM and QuICC
+        self.isEPM = False
+        
+        if file_type.lower()!='quicc':
+            self.isEPM = True
+        # assuming using QuICC state syntax
+        
+        # initialize the .parameters object
+        self.parameters = lambda: None
+                   
+        # hardcode set time and timestep
+        if not self.isEPM:
+            setattr(self.parameters, 'time', fin['run/time'][()])
+            setattr(self.parameters, 'timestep', fin['run/timestep'][()])
+            for at in fin['physical'].keys():
+                setattr(self.parameters, at, fin['physical'][at].value)
+        else:
+            setattr(self.parameters, 'time', fin['RunParameters/Time'][()])
+            setattr(self.parameters, 'timestep', fin['RunParameters/Step'][()])
+        
+        # import attributes
+
+        
+        #self.geometry = fin.attrs['type']
+        self.geometry = geometry.lower()
+        if(self.geometry not in list(['cartesian', 'shell', 'sphere'])):
+            raise RuntimeError("I'm sorry but we only deal with Cartesian, Shell and Sphere, try again later")
+        # geometry so far is
+    pass
 
 class SpectralState(BaseState):
     
@@ -858,9 +905,9 @@ class SpectralState(BaseState):
                 t_part = np.zeros_like(r)
 
             for n in range(self.specRes.N):
-                modeP_r=modeP_r+modeP[n]*wor.W(n, l, r)/r
-                dP = dP+modeP[n]*wor.diffW(n, l, r)
-                t_part = t_part + modeT[n]*wor.W(n,l,r)
+                modeP_r=modeP_r+modeP[n]*sphere.W(n, l, r)/r
+                dP = dP+modeP[n]*sphere.diffW(n, l, r)
+                t_part = t_part + modeT[n]*sphere.W(n,l,r)
 
             q_part =  modeP_r*l*(l+1)#same stuff
             s_part = modeP_r + dP 
@@ -1093,4 +1140,150 @@ class SpectralState(BaseState):
             pol_product /= (2*vol)
         return tor_product + pol_product
             
- 
+class PhysicalState(BaseState):
+    
+    
+    def __init__(self, filename, geometry, file_type='QuICC'):
+        
+        # apply the read of the base class
+        BaseState.__init__(self, filename, geometry, file_type='QuICC')
+        fin = self.fin
+        # read the grid
+        if not self.isEPM:
+            for at in fin['mesh']:
+                setattr(self, at, fin['mesh'][at].value)
+            
+            # find the fields
+            self.fields = lambda:None
+            #temp = object()
+            for group in fin.keys():
+
+                for subg in fin[group]:
+
+                    # we check to import fields which are at least 2-dimensional tensors
+                    field = np.array(fin[group][subg])
+                    if len(field.shape)<2:
+                        continue
+
+                    # set attributes
+                    setattr(self.fields, subg, field)
+                    
+        else:
+            for at in fin['FSOuter/grid']:
+                setattr(self, at, fin['FSOuter/grid'][at].value)
+            
+            # find the fields
+            self.fields = lambda:None
+            #temp = object()
+            for group in fin['FSOuter'].keys():
+
+                for subg in fin['FSOuter'][group]:
+
+                    # we check to import fields which are at least 2-dimensional tensors
+                    field = np.array(fin['FSOuter'][group][subg])
+                    if len(field.shape)<2:
+                        continue
+
+                    # set attributes
+                    setattr(self.fields, subg, field)
+
+
+    # define function for equatorial plane visualization from the visState0000.hdf5 file
+    def makeMeridionalSlice(self, phi=None, fieldname = 'velocity'):
+
+        # some parameters just in case
+        eta = self.parameters.rratio
+        ri = eta/(1-eta)
+        ro = 1/(1-eta)
+        a, b = .5, .5*(1+eta)/(1-eta)
+
+        # find the grid in radial and meridional direction
+        r = self.grid_r #.value[idx_r]
+        theta =  self.grid_theta
+        
+        rr, ttheta = np.meshgrid(self.grid_r, self.grid_theta)
+        X = rr*np.sin(ttheta)
+        Y = rr*np.cos(ttheta)
+
+        if phi == None:
+            # select the 0 meridional plane value for 
+            idx_phi0 = (self.grid_phi==0)
+            
+            Field1 = np.mean(getattr(self.fields, fieldname+'_r')[:, :, idx_phi0], axis=2)
+            Field2 = np.mean(getattr(self.fields, fieldname+'_theta')[:, :, idx_phi0], axis=2)
+            Field3 = np.mean(getattr(self.fields, fieldname+'_phi')[:, :, idx_phi0], axis=2)
+        else:
+            phi = self.grid_phi
+            Field1 = interp1d(phi, getattr(self.fields, fieldname+'_r'), axis=2)(phi)
+            Field2 = interp1d(phi, getattr(self.fields, fieldname+'_theta'), axis=2)(phi)
+            Field3 = interp1d(phi, getattr(self.fields, fieldname+'_phi'), axis=2)(phi)
+            
+        return X, Y, [Field1, Field2, Field3]
+
+
+    
+    # define function for equatorial plane visualization from the visState0000.hdf5 file
+    def makeEquatorialSlice(self, fieldname = 'velocity'):
+                        
+        # select the r grid in the bulk of the flow
+        #idx_r = (fopen['mesh/grid_r']>ri+delta) & (fopen['mesh/grid_r']<ro-delta)
+        #idx_r = (fopen['mesh/grid_r'].value>0.)
+        
+        # select the 0 meridional plane value for
+        theta_grid = self.grid_theta
+        neq = int(len(theta_grid)/2)
+        if len(theta_grid) % 2 == 0:
+            idx_theta = (theta_grid == theta_grid[neq-1]) | (theta_grid == theta_grid[neq])
+            print(theta_grid[neq-1]-np.pi/2, theta_grid[neq]-np.pi/2)
+        else:
+            idx_theta = (theta_grid == theta_grid[neq])
+            
+        # find the grid in radial and meridional direction
+        #r = fopen['mesh/grid_r'].value[idx_r]
+        r = self.grid_r
+        phi = self.grid_phi
+        
+        rr, pphi = np.meshgrid(r, phi)
+        # compute the values for x and y
+        X = np.cos(pphi)*rr
+        Y = np.sin(pphi)*rr
+        
+        Field1 = np.mean(getattr(self.fields, fieldname+'_r')[:,idx_theta,:], axis=1)
+        Field2 = np.mean(getattr(self.fields, fieldname+'_theta')[:,idx_theta,:], axis=1)
+        Field3 = np.mean(getattr(self.fields, fieldname+'_phi')[:,idx_theta,:], axis=1)
+
+        return X, Y, [Field1, Field2, Field3]
+        
+    # define function for equatorial plane visualization from the visState0000.hdf5 file
+    def makeIsoradiusSlice(self, r=None, fieldname = 'velocity'):
+        
+        # some parameters just in case
+        eta = self.parameters.rratio
+        ri = eta/(1-eta)
+        ro = 1/(1-eta)
+        a, b = .5, .5*(1+eta)/(1-eta)
+
+         # find the grid in radial and meridional direction
+        theta = self.grid_theta
+        phi = self.grid_phi
+        
+        TTheta, PPhi = np.meshgrid(theta, phi)
+        r_grid = self.grid_r
+        if r==None:
+            # select the 0 meridional plane value for
+            neq = int(len(r_grid)/2)
+            if len(r_grid) % 2 == 0:
+                idx_r = (r_grid == r_grid[neq-1]) | (r_grid == r_grid[neq])
+            else:
+                idx_theta = (theta_grid == theta_grid[neq])
+            
+            Field1 = np.mean(getattr(self.fields, fieldname+'_r')[idx_r, :, :], axis=0)
+            Field2 = np.mean(getattr(self.fields, fieldname+'_theta')[idx_r, :, :], axis=0)
+            Field3 = np.mean(getattr(self.fields, fieldname+'_phi')[idx_r, :, :], axis=0)
+        else:
+
+            Field1 = interp1d(r_grid, getattr(self.fields, fieldname+'_r'), axis=0)(r)
+            Field2 = interp1d(r_grid, getattr(self.fields, fieldname+'_theta'), axis=0)(r)
+            Field3 = interp1d(r_grid, getattr(self.fields, fieldname+'_phi'), axis=0)(r)
+                    
+        return TTheta, PPhi, [Field1, Field2, Field3]
